created Orm superclass, moved redundant methods from Raw and Product classes to Orm class
